feat(fs): add symlink_metadata and File::seek; unify Metadata::mode()
Close the remaining fs semantic gaps (release plan 3.8; the SO_REUSEADDR piece
landed separately):

- symlink_metadata(path): reports the link itself instead of following it, so
  Metadata::is_symlink() can be true. The backend already threaded a
  follow_symlinks flag; this just passes false.
- File::seek(SeekFrom) -> u64: repositions the shared kernel fd cursor via an
  inline lseek(2). Both backends drive sequential read/write off that cursor, so
  seek composes with them.
- Metadata::mode(): macOS masked to 0o7777 (permission bits only) while Linux
  returned the full st_mode. Both now return the full st_mode (file-type +
  permission bits) widened to u32, matching std::os::unix::fs::MetadataExt::mode.

Tests in tests/fs_metadata.rs. Public API snapshot updated.

+++ src/op/fs.rs
@@ -38,7 +38,10 @@ pub enum FileType {
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct RawMetadata {
     pub file_type: FileType,
-    pub mode: u16,
+    /// Full POSIX `st_mode` (file-type bits and permission bits), matching
+    /// `std::os::unix::fs::MetadataExt::mode`. `u32` for parity with std even
+    /// though the current backends fit it in 16 bits.
+    pub mode: u32,
     pub len: u64,
 }
